# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/karyawan/absensi": {
            "get": {
                "tags": [
                    "AbsensiKaryawan"
                ],
                "summary": "Ambil data absensi karyawan, dengan pencarian berdasarkan nama karyawan (wajib)",
                "operationId": "GetAbsensi",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian nama karyawan (wajib diisi)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data absensi karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data absensi karyawan tersedia"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "nip": {
                                                        "type": "string",
                                                        "example": "6535534"
                                                    },
                                                    "tanggal": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-06-01"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "Hadir",
                                                            "Izin",
                                                            "Sakit",
                                                            "Alpha"
                                                        ],
                                                        "example": "Hadir"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-02T10:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-06-02T11:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawan/absensi/create": {
            "post": {
                "tags": [
                    "AbsensiKaryawan"
                ],
                "summary": "Tambah data absensi karyawan",
                "operationId": "CreateAbsensiKaryawan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_karyawan",
                                    "tanggal",
                                    "status"
                                ],
                                "properties": {
                                    "id_karyawan": {
                                        "description": "ID karyawan yang melakukan absensi",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tanggal": {
                                        "description": "Tanggal absensi (format YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-02"
                                    },
                                    "status": {
                                        "description": "Status absensi (Hadir, Izin, Sakit, Alpha)",
                                        "type": "string",
                                        "example": "Hadir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Absensi karyawan berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi karyawan berhasil ditambahkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal atau absensi sudah ada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawan/absensi/{id}/update": {
            "put": {
                "tags": [
                    "AbsensiKaryawan"
                ],
                "summary": "Update data absensi karyawan berdasarkan ID absensi",
                "operationId": "UpdateAbsensiById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID absensi yang akan diupdate",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Status absensi (Hadir, Izin, Sakit, Alpha)",
                                        "type": "string",
                                        "example": "Hadir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Absensi karyawan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi karyawan berhasil diupdate"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data absensi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data absensi tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawan/absensi/{id}/delete": {
            "delete": {
                "tags": [
                    "AbsensiKaryawan"
                ],
                "summary": "Hapus data absensi karyawan berdasarkan ID absensi",
                "operationId": "DeleteAbsensiById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID absensi yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data absensi berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data absensi berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data absensi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data absensi tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan saat menghapus data absensi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Gagal menghapus data absensi: ..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "security": [
            {
                "bearerAuth": []
            }
        ]
    }
}